Move cublas_handle into own header file
